Created HomeViewModel and Refactored the App to MVVM
