+++ Home/Models/Activity.swift
@@ -1,0 +1,17 @@
+//
+//  Activity.swift
+//  FitnessApp
+//
+//  Created by Jordan Austin on 12/29/23.
+//
+
+import SwiftUI
+
+struct Activity {
+    let id : Int
+    let title : String
+    let subtitle : String
+    let image : String
+    let tintColor : Color
+    let amount : String
+}
