+++ Home/Models/Workout.swift
@@ -1,0 +1,18 @@
+//
+//  Workout.swift
+//  FitnessApp
+//
+//  Created by Jordan Austin on 12/29/23.
+//
+
+import SwiftUI
+
+struct Workout {
+    let id : Int
+    let title : String
+    let image : String
+    let duration : String
+    let colorTint : Color
+    let date : String
+    let calories : String
+}
