+++ Home/Views/WorkoutCard.swift
@@ -7,15 +7,6 @@
 
 import SwiftUI
 
-struct Workout {
-    let id : Int
-    let title : String
-    let image : String
-    let duration : String
-    let colorTint : Color
-    let date : String
-    let calories : String
-}
 struct WorkoutCard: View {
     @State var workout : Workout
